+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## V 1.4.1 : XXX Update - XX/07/19
 
+    - Color : Add parameter to apply darker or lighter x times
+    
     - Bug Fix : Sprite is rescale at every change of sprite in SpriteComponent
     - Bug Fix : If tileset isn't in the same director than tilemap, sprites isn't found
     - Bug Fix : ControlComponent save the keypressed when change world